logik und ui auftrennen für authentication
